Refactor: 설정 탭 - 프로필 이미지 수정 화면 Base 코드 적용
