+++ meaning-out/Controllers/Setting/EditProfileImageViewController.swift
@@ -7,13 +7,9 @@
 
 import UIKit
 
-final class EditProfileImageViewController: UIViewController {
+final class EditProfileImageViewController: BaseViewController {
     
-    let profileImageView = UIView()
-    let profileImage = UIImageView()
-    
-    let cameraImageView = UIView()
-    let cameraImage = UIImageView()
+    private let editProfileImageView = EditProfileImageView()
     
     private var isDefaultSelected = true
     
@@ -24,108 +20,27 @@ final class EditProfileImageViewController: UIViewController {
         }
     }
     
-    // 프로필 이미지 컬렉션 뷰
-    lazy var profileCollectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionviewLayout())
-    
-    // 프로필 이미지 컬렉션 뷰 레이아웃 구성
-    func collectionviewLayout() -> UICollectionViewLayout {
-        let layout = UICollectionViewFlowLayout()
-        
-        let sectionSpaciing: CGFloat = 16
-        let cellSpacing: CGFloat = 16
-        
-        let width = UIScreen.main.bounds.width - (sectionSpaciing) - (cellSpacing)
-        layout.itemSize = CGSize(width: width / 6, height: width / 6)
-        layout.scrollDirection = .vertical
-        
-        layout.minimumInteritemSpacing = cellSpacing
-        layout.minimumLineSpacing = cellSpacing
-        layout.sectionInset = UIEdgeInsets(top: sectionSpaciing, left: sectionSpaciing, bottom: sectionSpaciing, right: sectionSpaciing)  // 셀과 뷰 사이 간격
-        
-        return layout
+    override func loadView() {
+        self.view = editProfileImageView
     }
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        configureView()
-        configureHierarchy()
-        configureLayout()
-        configureUI()
         configureData()
     }
     
-    private func configureView() {
-        view.backgroundColor = Resource.Colors.white
+    override func configureViewController() {
         navigationItem.title = Constants.Title.edit.rawValue
-        
         addImgBarBtn(image: Resource.SystemImages.left, style: .plain, target: self, action: #selector(popViewController), type: .left)
-    }
-    
-    private func configureHierarchy() {
-        cameraImageView.addSubview(cameraImage)
         
-        profileImageView.addSubview(profileImage)
-        profileImageView.addSubview(cameraImageView)
-    
-        view.addSubview(profileImageView)
-        view.addSubview(profileCollectionView)
-        
-        profileCollectionView.delegate = self
-        profileCollectionView.dataSource = self
-        profileCollectionView.register(ProfileImageCollectionViewCell.self, forCellWithReuseIdentifier: ProfileImageCollectionViewCell.id)
-    }
-    
-    private func configureLayout() {
-        profileImageView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(16)
-            $0.size.equalTo(100)
-            $0.centerX.equalTo(view.safeAreaLayoutGuide)
-        }
-        
-        profileImage.snp.makeConstraints {
-            $0.top.bottom.equalTo(profileImageView)
-            $0.size.equalTo(profileImageView)
-            $0.centerX.equalTo(profileImageView)
-        }
-        
-        cameraImageView.snp.makeConstraints {
-            $0.trailing.equalTo(profileImage)
-            $0.bottom.equalTo(profileImage).inset(8)
-            $0.size.equalTo(24)
-        }
-        
-        cameraImage.snp.makeConstraints {
-            $0.center.equalTo(cameraImageView)
-            $0.size.equalTo(15)
-        }
-        
-        profileCollectionView.snp.makeConstraints {
-            $0.top.equalTo(profileImageView.snp.bottom).offset(32)
-            $0.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(16)
-            $0.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-    }
-    
-    private func configureUI() {
-        profileImage.backgroundColor = Resource.Colors.white
-        profileImage.clipsToBounds = true
-        profileImage.layer.cornerRadius = 50  // 추후 상수화 하기
-        profileImage.layer.borderColor = Resource.Colors.primary.cgColor
-        profileImage.layer.borderWidth = CGFloat(Constants.Integer.borderWidth.rawValue)
-        profileImage.contentMode = .scaleAspectFit
-        
-        cameraImageView.backgroundColor = Resource.Colors.primary
-        cameraImageView.layer.cornerRadius = 12  // 추후 상수화 하기
-        
-        cameraImage.image = Resource.SystemImages.camara
-        cameraImage.contentMode = .scaleAspectFit
-        cameraImage.tintColor = Resource.Colors.white
+        editProfileImageView.profileCollectionView.delegate = self
+        editProfileImageView.profileCollectionView.dataSource = self
+        editProfileImageView.profileCollectionView.register(ProfileImageCollectionViewCell.self, forCellWithReuseIdentifier: ProfileImageCollectionViewCell.id)
     }
     
     private func configureData() {
         // 프로필 랜덤 노출 -> 가입 시 고정값으로 수정
-        profileImage.image = Resource.Images.profiles[profileNum]
+        editProfileImageView.profileImage.image = Resource.Images.profiles[profileNum]
     }
 
 }
@@ -160,6 +75,6 @@ extension EditProfileImageViewController: UICollectionViewDelegate, UICollection
         profileNum = indexPath.item
         UserDefaultsManager.profile = profileNum
         isDefaultSelected = false
-        profileImage.image = Resource.Images.profiles[profileNum]
+        editProfileImageView.profileImage.image = Resource.Images.profiles[profileNum]
     }
 }
